feat: generate semver tags
- SemVer tags are :major, :major.minor and :major.minor.patch.
- They will be generated by default if all the criterions match:
  - version is a valid semver
  - `--latest` is enabled
- New option: --tag-semver, if set, it will enforce the generation of
semver tags, and will cause docker-meta to fail if a non-valid semver is
supplied as `version`.

BREAKING CHANGE: Now semver tags are generated by default if a valid
semver is detected as `version` (you can disable this behavior with the
new option `--no-tag-semver`).

BREAKING CHANGE: Now the version will be normalized if it is a valid
semver, for example: `v1.0.0` becomes `1.0.0`.